Feat: table is now a feature
